removed EPS reader from the compiled readers
A part of this reader is written in python which forces us to package python or to require the users to
manually install python. For now the solution is to simply remove this reader since it is rarely used.
If we need the reader back, it is probably simpler to translate it to c++ and avoid problems